moving statusbar example to examples
